fix: normalize section class key

+++ EduOS.Core/Entities/Academic/Section.cs
@@ -4,7 +4,7 @@ namespace EduOS.Core.Entities.Academic
 {
     public class Section : BaseTenantEntity
     {
-        public int ClassId { get; set; }
+        public long ClassId { get; set; }
         public string Name { get; set; } = string.Empty; // A/B/C
         public int Capacity { get; set; }
         public bool IsActive { get; set; } = true;
